LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/local-if-substitution-expected.txt


FAIL var() in if() condition's custom property value substitutes locally assert_equals: expected "PASS" but got "--f()"
FAIL var() in if() condition's specified value substitutes locally assert_equals: expected "PASS" but got "--f()"
FAIL var() in if() declaration value substitutes locally assert_equals: expected "PASS" but got "--f()"
FAIL var() in if() condition's custom property value substitutes locally, argument assert_equals: expected "PASS" but got "--f(3px)"
FAIL var() in if() condition's specified value substitutes locally, argument assert_equals: expected "PASS" but got "--f(3px)"
FAIL var() in if() declaration value substitutes locally, argument assert_equals: expected "PASS" but got "--f(PASS)"
FAIL dashed function in if() declaration value assert_equals: expected "PASS" but got "--f()"
FAIL dashed function with argument in if() declaration value assert_equals: expected "PASS" but got "--f(PASS)"
FAIL if() cycle through local assert_equals: expected "PASS" but got "--f()"
FAIL if() cycle in condition custom property through local assert_equals: expected "PASS" but got "--f()"
FAIL if() cycle in condition specified value through local assert_equals: expected "PASS" but got "--f()"
FAIL if() cycle through function assert_equals: expected "PASS" but got "--f()"
FAIL if() no cycle in overridden local assert_equals: expected "PASS" but got "--f()"
FAIL if() no cycle in overridden argument assert_equals: expected "PASS" but got "--f(3px)"
FAIL CSS-wide keywords are interpreted locally (initial) assert_equals: expected "PASS" but got "--g()"
FAIL CSS-wide keywords are interpreted locally (inherit) assert_equals: expected "PASS" but got "--g()"
FAIL CSS-wide keywords are interpreted locally (guaranteed-invalid, initial) assert_equals: expected "PASS" but got "--f()"
FAIL CSS-wide keywords are interpreted locally (guaranteed-invalid, unset) assert_equals: expected "PASS" but got "--f()"
FAIL CSS-wide keywords are interpreted locally (revert) assert_equals: expected "PASS" but got "--f()"
FAIL CSS-wide keywords are interpreted locally (revert-layer) assert_equals: expected "PASS" but got "--f()"
